third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-async/async-callstack-set-interval.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/debugger-test.js"></script>
<script>
var intervalId;
var count = 0;

function testFunction()
{
    intervalId = setInterval(callback, 0);
}

function callback()
{
    if (count === 0) {
        debugger;
    } else if (count === 1) {
        debugger;
    } else {
        clearInterval(intervalId);
        debugger;
    }
    ++count;
}

function test() {
  var totalDebuggerStatements = 3;
  SourcesTestRunner.runAsyncCallStacksTest(totalDebuggerStatements);
}

</script>
</head>

<body onload="runTest()">
<p>
Tests asynchronous call stacks for setInterval.
</p>
</body>
</html>
